Move loop signal handling to main module and simplify code
Signal masking (sigprocmask) must be done before any threads are created for
the masking to be inherited by all threads. Otherwise, signals are captured
by unmasked threads and the default system handlers are used instead.

Therefore, move the initialisation of signals and handlers to the main module
where it can be done before any other threads are created, e.g. platform,
video and connection threads.

Also, this allows to simplify the loop module code.
